Put smart account scripts in own service.
